Create Introduction to Data Projects  - Titanic Competition
This file is the first project done for the "Data Visualization with Python" course taught by Jose Portilla on Udemy. The project takes data about passengers of the Titanic and analyzes what factors affected a passenger's survival rate. It also characterizes the passengers through various metrics. Most of the work was a step-by-step, assisted walkthrough that followed along with the instructor, with a few deprecated functions requiring a deep dive into documentation for Seaborn and pandas.