Converting .txt files to .ipynb files
I've found this to be a quick and easy way to convert .txt files of Juptyer Notebook code (.ipynb files). It does seem to have deprecated, but until it is rendered obsolete in a later version, this is a way to download these files, convert them to an interactable file type.