Job Changes for Data Scientists - HR Analytics
This was a project with data that contains generic information about currently employed data scientists and whether or not they were looking for another job. There's some really interesting trends that are uncovered during the analysis of this dataset. Try out different combinations of comparisons to find more trends that highlight important characteristics about this group of data scientists. The determination of the probabilities at the end of the notebook (i.e. of how likely some random person will look for another job) is incorrect and will be corrected in the next few days (some time around June 26th-June 28th, 2021). Feel free to conduct your own statistical analysis at the end of the notebook and see if you can discover something fascinating from the probability distributions.