Trends in Funding for US Solar Projects
This Juptyer notebook contains data visualizations of projects related to the US's Solar Energy Industry and trends of how the funding for these projects were distributed. This project mainly focuses on active projects. Factors considered include (but aren't limited to) what types of projects the US government tends to fund, how much of the costs they fund for a project, and which states are funded more often.